исправлено наименование модулей, общие константы вынесены в отдельный модуль
