--- src/utils/hidden-caption.js
@@ -1,7 +1,0 @@
-const createHiddenCaptionTemplate = (title) => {
-  return (
-    `<h2 class="visually-hidden">${title}</h2>`
-  );
-};
-
-export {createHiddenCaptionTemplate};
