+++ src/view/general-constants.js
@@ -1,0 +1,12 @@
+const EventMessage = {
+  NO_EVENTS: `Click New Event to create your first point`,
+  LOADING: `Loading...`,
+};
+
+const BlockTitle = {
+  SWITCH: `Switch trip view`,
+  FILTER: `Filter events`,
+  TRIP_EVENTS: `Trip events`
+};
+
+export {EventMessage, BlockTitle};
