+++ src/view/trip-controls.js
