+++ src/view/trip-day.js
