+++ src/view/trip-event-button.js
@@ -1,7 +1,7 @@
-const createTripEventBtnTemplate = () => {
+const createTripEventButtonTemplate = () => {
   return (
     `<button class="trip-main__event-add-btn  btn  btn--big  btn--yellow" type="button">New event</button>`
   );
 };
 
-export {createTripEventBtnTemplate};
+export {createTripEventButtonTemplate};
