+++ src/view/trip-event.js
