+++ src/view/trip-events.js
