+++ src/view/trip-info.js
